Refactoring: change project structure
