+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require('./get-all');
+const getById = require('./get-by-id');
+const add = require('./add');
+const updateById = require('./update-by-id');
+const deleteById = require('./delete-by-id');
+
+module.exports = {
+  getAll,
+  getById,
+  add,
+  updateById,
+  deleteById,
+};
